A Termodinâmica entra antes da Cosmologia, e o par fecha na unidade.
O rendimento máximo sai das duas integrais fechadas — ∮dU=0 dá W=Q_q−Q_f,
∮dQ/T=0 dá Q_f/Q_q=T_f/T_q — e o tecto η=1−T_f/T_q não é limite de
engenharia: é o ciclo fechar. Junto vai o veredicto NEGATIVO sobre a razão
focal: (1−e)/(1+e) dá um T_f/T_q válido, e é por isso que não serve —
qualquer bijeção de (0,1) daria. Uma parametrização não prova nada porque
não pode falhar.

O piso do involutivo é zero EXACTO, e a prova é a fibra: apagar um bit é ter
dois estados com a mesma saída, isto é G>1; uma involução é bijetiva, logo
G≡1, logo a soma é vazia. Dissipar é dobrar. E o par dual fecha na unidade,
S_negro·S_branco=1 com log-soma 0 — donde «a entropia cresce» ser lei de
METADE do par: a soma não se move.

A ponte termo↔cosmo fica escrita como está: a mesma FORMA não é a mesma LEI,
e o que a decidiria é exibir a construção, não emparelhar — pelo mesmo
critério que fechou o Carnot.

tests/dissipa.c: o custo por segundo saía −2147483648 num `long` de 32 bits.
Não era asserido, logo nenhum bloco o apanhava. Terceira ocorrência do mesmo
padrão nesta sessão; o tecto diz-se agora no tipo.

+++ tests/dissipa.c
@@ -129,11 +129,14 @@ int main(void){
          *   k_B·T·ln2 = 2 870 978 rJ   (resto 0,885… de rJ, truncado — dito)
          *
          * ≈ 2,871·10⁻²¹ J por bit apagado, que é o limite de Landauer a 300 K. */
-        const long KT_LN2_rJ = 2870978L;             /* 10⁻²⁷ J; ln2 truncado, ver acima */
-        printf("      a %ld rJ (10⁻²⁷ J) por bit: destrutiva %ld rJ, involutiva %ld rJ\n",
-               KT_LN2_rJ, perdidos_d*KT_LN2_rJ, perdidos_i*KT_LN2_rJ);
-        printf("      e uma DRAM que refresque isto 8192x/s pagaria %ld rJ/s so' por existir\n",
-               (long)N*32*8192*KT_LN2_rJ);
+        /* 64 BITS EXPLICITOS: N*32*8192*KT_LN2_rJ passa dos 2^31, e num alvo onde o
+         * `long` tem 32 bits saia' NEGATIVO — um numero publicado errado, e este nao
+         * e' asserido, logo nenhum bloco o apanhava. O tecto diz-se no tipo. */
+        const long long KT_LN2_rJ = 2870978LL;       /* 10⁻²⁷ J; ln2 truncado, ver acima */
+        printf("      a %lld rJ (10⁻²⁷ J) por bit: destrutiva %lld rJ, involutiva %lld rJ\n",
+               KT_LN2_rJ, (long long)perdidos_d*KT_LN2_rJ, (long long)perdidos_i*KT_LN2_rJ);
+        printf("      e uma DRAM que refresque isto 8192x/s pagaria %lld rJ/s so' por existir\n",
+               (long long)N*32*8192*KT_LN2_rJ);
         ok("E O PISO DA INVOLUTIVA E' ZERO EXACTO, e nao um numero pequeno: nao ha' bit"
            " apagado, logo nao ha' minimo termodinamico nenhum a pagar. E os dois lados"
            " têm de ser VISÍVEIS para a frase valer: a destrutiva paga um número > 0 na"
